Fix header styling
