+++ extension/src/content-script/components/AgentView.tsx
@@ -12,7 +12,7 @@ export const AgentView = () => {
   const [currentView, setCurrentView] = useState<AGENT_VIEWS>(AGENT_VIEWS.ACTIVE_AGENTS)
 
   return (
-    <div className="p-6  text-white bg-black w-[500px] absolute right-0 top-10 h-[800px] overflow-scroll">
+    <div className="p-6 text-white bg-black w-[500px] fixed right-0 top-10 h-[800px] z-50">
       {currentView === AGENT_VIEWS.ACTIVE_AGENTS && (
         <ActiveAgents setCurrentView={setCurrentView} />
       )}
